separated into .c and .h
